requirements.txt
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
